+++ spec2vec/__init__.py
@@ -1,3 +1,4 @@
+from . import serialization
 from .__version__ import __version__
 from .Document import Document
 from .logging_functions import _init_logger
@@ -13,6 +14,7 @@ __all__ = [
     "__version__",
     "calc_vector",
     "Document",
+    "serialization",
     "SpectrumDocument",
     "Spec2Vec",
 ]